key action now takes two parameters: source and destination id - key table now use action table implementation - ui_area: can now respond to draw event - window manager now tells ui area root went to draw

+++ include/ui/action_table.h
@@ -7,6 +7,6 @@
   int Action_table_create(void);
   void Action_table_allocate_slots(const short table_id, const short count);
   void Action_table_set_action(const short table_id, const short slot_id, ActionFn action);
-  void Action_table_execute(const short table_id, const short slot_id, const uint32_t source_id, const uint32_t destination_id);
+  void Action_table_execute(const short table_id, const short slot_id, const short source_id, const short destination_id);
 
 #endif
